# openapi.yaml
openapi: 3.1.0
info:
  title: Piano Lesson Booking API
  version: 1.0.0
  description: |
    The HTTP API expected by the fixture-backed client in `frontend/`.

    Teacher endpoints require an authenticated admin session. Student endpoints
    do not use the admin session: the unguessable token in the URL is the
    student's credential and must not be logged or exposed to another student.

    Dates and times are evaluated in the studio's configured time zone. Date-time
    values are RFC 3339 strings with an offset. The server, rather than the
    frontend, determines availability, day locks, and whether a lesson can move.
servers:
  - url: /
    description: Same origin as the frontend
...
paths:
...
  /s/{token}:
    get:
      tags: [Student]
      summary: Get the student's lesson page data
      operationId: getStudentView
      security: []
      parameters:
        - $ref: "#/components/parameters/StudentToken"
      responses:
        "200":
          description: Student, package, and upcoming lesson data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentView"
        "404":
          $ref: "#/components/responses/InvalidToken"
  /s/{token}/slots:
    get:
      tags: [Student]
      summary: List slots available for rescheduling
      description: |
        Returns only bookable slots. Slots outside the move window, locked days,
        blackouts, and occupied slots are omitted. When `lessonId` is supplied,
        the lesson's current slot is excluded from the occupied-slot calculation.
      operationId: getStudentOpenSlots
      security: []
      parameters:
        - $ref: "#/components/parameters/StudentToken"
        - name: from
          in: query
          required: true
          description: First local calendar date to consider, inclusive.
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          description: Last local calendar date to consider, inclusive.
          schema:
            type: string
            format: date
        - name: lessonId
          in: query
          required: false
          description: Lesson being moved.
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: Available slots sorted by start time
          content:
            application/json:
              schema:
                type: object
                required: [slots]
                properties:
                  slots:
                    type: array
                    items:
                      $ref: "#/components/schemas/OpenSlot"
        "404":
          $ref: "#/components/responses/InvalidToken"
  /s/{token}/lessons/{lessonId}/reschedule:
    post:
      tags: [Student]
      summary: Move a lesson to an open slot
      description: |
        The server must atomically verify that the destination remains open.
        A stale selection returns `SLOT_TAKEN`; the frontend keeps the picker
        open and requests the slot list again.
      operationId: rescheduleStudentLesson
      security: []
      parameters:
        - $ref: "#/components/parameters/StudentToken"
        - $ref: "#/components/parameters/LessonId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [startsAt]
              properties:
                startsAt:
                  type: string
                  format: date-time
                  examples: ["2026-09-17T15:15:00+02:00"]
      responses:
        "200":
          description: Lesson moved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonMutationSuccess"
        "400":
          description: The move is outside the allowed window.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: The original day is locked, or the student is paused or flagged.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          $ref: "#/components/responses/InvalidToken"
        "409":
          description: The destination slot is no longer available.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /s/{token}/pause-request:
    post:
      tags: [Student]
      summary: Ask for a break and release the weekly slot
      operationId: requestStudentPause
      security: []
      parameters:
        - $ref: "#/components/parameters/StudentToken"
      responses:
        "200":
          description: Student paused and teacher notified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentMutationSuccess"
        "403":
          description: The student cannot request a pause in the current state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          $ref: "#/components/responses/InvalidToken"
...
  /admin/students/{studentId}:
    get:
      tags: [Admin students]
      summary: Get a student and lesson history
      operationId: getAdminStudent
      security:
        - adminSession: []
      parameters:
        - $ref: "#/components/parameters/StudentId"
      responses:
        "200":
          description: Student detail
          content:
            application/json:
              schema:
                type: object
                required: [student, lessons]
                properties:
                  student:
                    $ref: "#/components/schemas/AdminStudent"
                  lessons:
                    type: array
                    items:
                      $ref: "#/components/schemas/Lesson"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/MutationError"
...
components:
  securitySchemes:
    adminSession:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token returned by `POST /auth/token`.
  parameters:
    StudentToken:
      name: token
      in: path
      required: true
      description: Unguessable student access token; possession grants access to this student only.
      schema:
        type: string
        minLength: 1
    StudentId:
      name: studentId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    LessonId:
      name: lessonId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
    InvalidToken:
      description: The student token is not valid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            error:
              code: INVALID_TOKEN
              message: This link isn't valid. Ask your teacher for a new one.
    Unauthenticated:
      description: A valid admin session is required.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AuthErrorResponse"
...
    MutationError:
      description: The mutation could not be completed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AdminErrorResponse"
  schemas:
    StudentStatus:
      type: string
      enum: [active, paused, flagged]
    LessonStatus:
      type: string
      enum: [scheduled, done]
    ApiErrorCode:
      type: string
      enum:
        - SLOT_TAKEN
        - DAY_LOCKED
        - OUTSIDE_WINDOW
        - STUDENT_PAUSED
        - STUDENT_FLAGGED
        - INVALID_TOKEN
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          $ref: "#/components/schemas/ApiErrorCode"
        message:
          type: string
          minLength: 1
          description: User-facing wording owned by the server and rendered directly by the frontend.
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [ok, error]
      properties:
        ok:
          type: boolean
          const: false
        error:
          $ref: "#/components/schemas/ApiError"
    AuthErrorResponse:
      type: object
      additionalProperties: false
      required: [ok, error]
      properties:
        ok:
          type: boolean
          const: false
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              const: UNAUTHENTICATED
            message:
              type: string
    AdminErrorResponse:
      type: object
      additionalProperties: false
      required: [ok, error]
      properties:
        ok:
          type: boolean
          const: false
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              minLength: 1
            message:
              type: string
              minLength: 1
...
    StudentMutationSuccess:
      type: object
      additionalProperties: false
      required: [ok, data]
      properties:
        ok:
          type: boolean
          const: true
        data:
          $ref: "#/components/schemas/AdminStudent"
    LessonMutationSuccess:
      type: object
      additionalProperties: false
      required: [ok, data]
      properties:
        ok:
          type: boolean
          const: true
        data:
          $ref: "#/components/schemas/Lesson"
    Package:
      type: object
      additionalProperties: false
      required: [size, used, periodNo]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          description: Present in admin responses so the invoice mutation can address the package.
        size:
          type: integer
          enum: [5, 8, 10]
        used:
          type: integer
          minimum: 0
        periodNo:
          type: integer
          minimum: 1
    StudentSummary:
      type: object
      additionalProperties: false
      required: [id, name, status]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        name:
          type: string
        status:
          $ref: "#/components/schemas/StudentStatus"
    AdminStudent:
      type: object
      additionalProperties: false
      required: [id, name, status, token, slotDay, slotTime, pkg, invoiceSent]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        name:
          type: string
        status:
          $ref: "#/components/schemas/StudentStatus"
        token:
          type: string
          description: Personal-link token copied by the teacher UI.
        slotDay:
          $ref: "#/components/schemas/Weekday"
        slotTime:
          $ref: "#/components/schemas/LocalTime"
        pkg:
          $ref: "#/components/schemas/Package"
        invoiceSent:
          type: boolean
    StudentView:
      type: object
      additionalProperties: false
      required: [student, package, nextLesson, lessons, canRequestPause]
      properties:
        student:
          $ref: "#/components/schemas/StudentSummary"
        package:
          $ref: "#/components/schemas/Package"
        nextLesson:
          oneOf:
            - $ref: "#/components/schemas/StudentLesson"
            - type: "null"
        lessons:
          type: array
          description: Upcoming lessons sorted by `startsAt`, not by sequence number.
          items:
            $ref: "#/components/schemas/StudentLesson"
        canRequestPause:
          type: boolean
    Lesson:
      type: object
      additionalProperties: false
      required: [id, seq, studentId, startsAt, status]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        seq:
          type: integer
          minimum: 1
        studentId:
          type: integer
          format: int64
          minimum: 1
        startsAt:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/LessonStatus"
    StudentLesson:
      type: object
      additionalProperties: false
      required: [id, seq, startsAt, status, canMove]
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        seq:
          type: integer
          minimum: 1
        startsAt:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/LessonStatus"
        canMove:
          type: boolean
    OpenSlot:
      type: object
      additionalProperties: false
      required: [startsAt]
      properties:
        startsAt:
          type: string
          format: date-time
    Weekday:
      type: integer
      minimum: 1
      maximum: 5
      description: ISO-like weekday used by the frontend, where 1 is Monday and 5 is Friday.
    LocalTime:
      type: string
      pattern: "^(?:[01][0-9]|2[0-3]):[0-5][0-9]$"
      examples: ["16:15"]
...